add serde_typetree

+++ src/objects/pptr.rs
@@ -2,7 +2,7 @@ use serde::{Deserialize, Serialize};
 
 #[derive(Serialize, Deserialize, Debug, Clone)]
 pub struct PPtr {
-    pub m_FileID: i64,
+    pub m_FileID: i32,
     pub m_PathID: i64,
 }
